+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/diagnostics/DiagnosticCode.java
@@ -67,6 +67,8 @@ public enum DiagnosticCode {
     GRAPHQL_143,
     GRAPHQL_144,
     GRAPHQL_145,
+    GRAPHQL_146,
+    GRAPHQL_147,
     GRAPHQL_201,
     GRAPHQL_202,
     GRAPHQL_203,